removed matchData logic from generic csvFileReader
