+++ stats/src/CsvFileReader.ts
@@ -1,30 +1,17 @@
 import fs from 'fs';
 
-import { dateStringToDate } from './utils';
-import { MatchResult } from './MatchResults';
-
-type MatchData = [Date, string, string, number, number, MatchResult, string];
-
-export class CsvFileReader {
-  data: MatchData[] = [];
+export abstract class CsvFileReader {
+  data: any = [];
 
   constructor(public filename: string) {}
+
+  abstract mapRow(row: string[]): /* MatchData */ any;
 
   read(): void {
     this.data = fs
       .readFileSync(this.filename, { encoding: 'utf-8' })
       .split('\n')
       .map((row: string): string[] => row.split(','))
-      .map(
-        (row: string[]): MatchData => [
-          dateStringToDate(row[0]),
-          row[1],
-          row[2],
-          parseInt(row[3]),
-          parseInt(row[4]),
-          row[5] as MatchResult,
-          row[6]
-        ]
-      );
+      .map(this.mapRow);
   }
 }
